Have vulkan copy commands take attachments instead of images.
The motivation for this is that with the new backend info structs the
sample count will be stored on non backend specific classes
(e.g. GrAttachment) and will not be part of the GrVkSurfaceSpec struct.
Thus the info will not be available from just a GrVkImage (or at least
until GrVkImage and GrVkAttachment get merged down the road).

This also removes some uses of GrVkImage::imageInfo() getter since that
struct will also be going away.

Bug: skia:12402 skia:10727
Change-Id: Ie90870e1d834e6ea6ba029d0b480641fb8d2cfa0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/446177
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>