Be more lenient with RLE BMPs
Bug: b/198566237

If the RLE data exceeds the width of the image, treat it as an incorrect
value in the number of pixels in the run. This allows decoding images
that are supported by other decoders (e.g. Mac Preview).

Add a test image to skimage tests.

Change-Id: I10955cbde3e28accf670417fa154a536c0732e6a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/488296
Reviewed-by: Derek Sollenberger <djsollen@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>