[fuzz] Require explicit SK_BUILD_FOR_*FUZZER
The existing define wasn't being used to compile :skia
(but it is for building the fuzzer code).

After this lands, oss-fuzz will be updated to set that
flag too. https://github.com/google/oss-fuzz/pull/6462

The change to SkTypes.h is to remove bit-rotted code that
prevents oss-fuzz from working.

Change-Id: I95b48d76dd3878d04ba3a5d021359ca200c20a0c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/449576
Reviewed-by: Michael Ludwig <michaelludwig@google.com>