Fix an issue reporting errors with check_deps.py
The script was previously checking to ensure that each entry was well
formatted, but then continued to try and parse it as if it was
regardless, causing an IndexError when trying to access the split array.

This change continues to the next iteration of the loop when an errors
is detected so that the script can complete normally and the errors can
be reported.

Change-Id: I830c74610ccce45e32a1afa679af7b0ee0881a4e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/451637
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Eric Boren <borenet@google.com>