fix rcp_precise for compiling with AVX
This will move highp AVX users from AVX to SSE41. This is about
7.5%[1] of users for software.

The underlying problem is that AVX works well with floating point,
it has little support for integer operations. So for lowp, AVX works
as SSE41 and in highp it works as HSW and SKX. The rcp_precise
has the distinction as being the only function that works as highp and
lowp. So, when compiled with -mavx it doesn't compile because in
highp it has 8 elements and in lowp it has 4 elements. So, move
it to be with SSE41 for highp.

[1] https://uma.googleplex.com/p/chrome/histograms?sid=ec095beddef4b2771a6a3f12a78f32c1

Bug: skia:12453

Change-Id: I2929312ef36cc0c0a68515ac74e5d953c0b7fd09
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/453716
Reviewed-by: Brian Osman <brianosman@google.com>