[skottie] Fill effect opacity workaround
The Fill effect has separate "color" and "opacity" components.

In AE, the color chooser does not allow modifying the alpha channel --
the only source for Fill transparency is the opacity property.

Thus, the alpha component of the color property should always be 1.

But in practice, it appears Bodymovin sometimes exports the color with a
zero alpha field (BM always encodes colors as 4-float arrays).

To workaround this issue, update Fill to ignore the color alpha and only
use the explicit opacity.

Change-Id: Ic4bbbc1ac91d255fe14e3261d1ae03340e053ce2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/454856
Commit-Queue: Florin Malita <fmalita@google.com>
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>